Migrates a project root helper out of the js API
